refactor: move validation utils
