+++ src/pipes/validation.pipe.ts
@@ -1,8 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import type { ArgumentMetadata, PipeTransform } from '@nestjs/common';
 
-import { getValidationSchema } from '../decorators/validation-schema.decorator.js';
-import { parseRequestBody } from '../utils/validation.utils.js';
+import { getValidationSchema, parseRequestBody } from '../utils/validation.utils.js';
 
 @Injectable()
 export class ValidationPipe implements PipeTransform {
